backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
pillow==10.1.0
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2 